Sugarcane_disease_detection/requirements.txt
matplotlib
torchvision
torch
tqdm
numpy
pillow
opencv-python
scikit-image
scikit-learn
pandas
transformers
timm
datasets
flask